pyproject.toml
[tool.poetry]
name = "sponsor_match"
version = "0.1.0"
description = "SponsorMatch AI: Intelligent sponsor–club matching web app"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
streamlit = "^1.15"
streamlit-folium = "^0.12.0"
folium = "^0.15.0"
SQLAlchemy = "^1.4"
PyMySQL = "^1.0"
pandas = "^1.3"
numpy = "^1.21"
scikit-learn = "1.5.0"
requests = "^2.26"
python-dotenv = "^0.19"

[tool.poetry.group.dev.dependencies]
pytest = "^6.2"
six = "^1.17.0"
lexicon = "^2.0.1"
invoke = "1.7.2"

[tool.poetry.scripts]
sponsor-match = "streamlit_app:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"